# backend/openapi.yaml
openapi: 3.0.3
info:
  title: IOC Backend API
  version: 1.0.0
  description: >
    API para registrar customers, usuários, emitir/autenticar tokens e gerenciar IOCs
    (globais e por customer). Backend em Node.js + Express com PostgreSQL.
servers:
  - url: /api
    description: Servidor local (relativo)
  - url: http://localhost:3000/api
    description: Desenvolvimento directo
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    adminKey:
      type: apiKey
      in: header
      name: X-Admin-Key
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
    CustomerRegisterRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          example: ACME SOC
    CustomerRegisterResponse:
      type: object
      properties:
        customerId:
          type: string
          format: uuid
        customerSecret:
          type: string
          description: Retornado apenas uma vez.
    UserRegisterRequest:
      type: object
      required: [customerId, customerSecret, email, password]
      properties:
        customerId:
          type: string
          format: uuid
        customerSecret:
          type: string
        email:
          type: string
          example: user@acme.com
        password:
          type: string
          minLength: 8
          example: "SenhaForte123!"
    UserRegisterResponse:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        customerId:
          type: string
          format: uuid
        email:
          type: string
    TokenCreateRequest:
      type: object
      required: [customerId, customerSecret]
      properties:
        customerId:
          type: string
          format: uuid
        customerSecret:
          type: string
    TokenCreateResponse:
      type: object
      properties:
        token:
          type: string
        tokenType:
          type: string
          example: Bearer
        expiresIn:
          type: string
          example: 12h
    TokenAuthRequest:
      type: object
      required: [token]
      properties:
        token:
          type: string
    TokenAuthResponse:
      type: object
      properties:
        valid:
          type: boolean
        payload:
          type: object
...
paths:
  /customer/register:
    post:
      tags: [Customer]
      summary: Registra um customer no backend com um customer ID único
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CustomerRegisterRequest" }
      responses:
        "201":
          description: Customer criado (retorna secret uma vez)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CustomerRegisterResponse" }
        "400":
          description: Requisição inválida
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /customer/user/register:
    post:
      tags: [Customer]
      summary: Registra um usuário do customer no backend baseado no customer ID
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UserRegisterRequest" }
      responses:
        "201":
          description: Usuário criado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserRegisterResponse" }
        "401":
          description: Secret inválido
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /customer/token/create:
    post:
      tags: [Customer]
      summary: Cria um token seguro (JWT) exclusivo para o customer ID
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TokenCreateRequest" }
      responses:
        "200":
          description: Token emitido
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TokenCreateResponse" }
        "401":
          description: Secret inválido
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /customer/token/auth:
    post:
      tags: [Customer]
      summary: Autentica o token do cliente (validação JWT)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TokenAuthRequest" }
      responses:
        "200":
          description: Token válido
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TokenAuthResponse" }
        "401":
          description: Token inválido/expirado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...